requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
PyMuPDF 
wikipedia
langchain-text-splitters
langchain-openai
Chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-chroma
langchain-ollama
streamlit
langchain-groq
fastapi
uvicorn
Langserve
langchain-core
sse_starlette